requirements.txt
Django==5.0.3
djangorestframework==3.15.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.2.0
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.7.0
boto3==1.34.84
django-storages==1.14.2
gunicorn==22.0.0
django-cors-headers==4.3.1